requirements.txt
Django>=3.2.4,<3.2.5
djangorestframework>=3.12.4,<3.12.5
psycopg2>=2.8.6,<2.8.7
Pillow>=8.2.0,<8.3.0

flake8>=3.9.2,<3.9.3